# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EthioMart E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, rating, popular]
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /sellers/register:
    post:
      operationId: registerSeller
      tags: [sellers]
      summary: Register as a seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSellerRequest"
      responses:
        "201":
          description: Seller application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"
...
  /sellers/me/analytics:
    get:
      operationId: getSellerAnalytics
      tags: [sellers]
      summary: Get seller analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Seller analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerAnalytics"
  /sellers/me/orders:
    get:
      operationId: getSellerOrders
      tags: [sellers]
      summary: Get seller's orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Seller orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a product review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Product reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        discount:
          type: integer
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        sold:
          type: integer
        featured:
          type: boolean
        status:
          type: string
          enum: [active, pending, rejected, inactive]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, price, stock, categoryId, sellerId, rating, reviewCount, status]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            specifications:
              type: object
              additionalProperties:
                type: string
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
            seller:
              $ref: "#/components/schemas/SellerSummary"
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        value:
          type: string
        price:
          type: number
        stock:
          type: integer
      required: [id, name, value, price, stock]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, totalPages]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        specifications:
          type: object
          additionalProperties:
            type: string
      required: [name, description, price, stock, categoryId]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        status:
          type: string
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: string
          enum: [telebirr, cbe_birr, cash_on_delivery]
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        trackingNumber:
          type: string
        estimatedDelivery:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orderNumber, userId, items, status, paymentStatus, subtotal, shipping, total, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required: [id, productId, quantity, price, subtotal]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, totalPages]
...
    Address:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        region:
          type: string
        country:
          type: string
          default: Ethiopia
        landmark:
          type: string
      required: [fullName, phone, street, city, region]
...
    RegisterSellerRequest:
      type: object
      properties:
        storeName:
          type: string
        storeDescription:
          type: string
        category:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        tinNumber:
          type: string
      required: [storeName, category, phone, address, city]
    Seller:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        storeName:
          type: string
        storeDescription:
          type: string
        logo:
          type: string
        category:
          type: string
        city:
          type: string
        rating:
          type: number
        totalSales:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, storeName, status]
    SellerSummary:
      type: object
      properties:
        id:
          type: integer
        storeName:
          type: string
        logo:
          type: string
        rating:
          type: number
        totalSales:
          type: integer
        city:
          type: string
      required: [id, storeName]
...
    SellerAnalytics:
      type: object
      properties:
        revenue:
          type: number
        orders:
          type: integer
        products:
          type: integer
        views:
          type: integer
        revenueChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        orderStatusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
      required: [revenue, orders, products, views]
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        images:
          type: array
          items:
            type: string
        verifiedPurchase:
          type: boolean
        helpful:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, productId, userId, userName, rating, verifiedPurchase, createdAt]
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
        averageRating:
          type: number
        ratingDistribution:
          type: object
          additionalProperties:
            type: integer
      required: [reviews, total, page, totalPages, averageRating]
    CreateReviewRequest:
      type: object
      properties:
        productId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        images:
          type: array
          items:
            type: string
      required: [productId, rating]
...
    ChartDataPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: string
      required: [date, value]
    TopProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
        sales:
          type: integer
        revenue:
          type: number
      required: [id, name, sales, revenue]
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [status, count]
...